fix(BUY-65474): cast LIMIT positional params as ::int in mcp.ts (#318)
Postgres prepared-statement path cannot infer the data type of a
positional parameter used in a LIMIT clause, raising:
  "could not determine data type of parameter \$3"
when the parameter is the last/only typed slot. This has been breaking
get_deals (fallback LIMIT \$3) and find_best_price (vector LIMIT \$3)
in production for the full BUY-64151 regression window.

Cast the three remaining untyped LIMIT params on main:
  - browse-mode LIMIT \$1 (search, ~L433)
  - get_deals regional fallback LIMIT \$3 (~L684)
  - find_best_price vector neighbor LIMIT \$3 (~L1231)

Existing LIMIT casts at L872/L877 unchanged. No behavioral change;
pure type annotation for the query planner.

Evidence: Fetch probe #208 2026-08-03T14:01Z — AC 0/4 with get_deals
\$3 error reproduced against deployed api.buywhere.ai/mcp.

+++ api/src/routes/mcp.ts
@@ -430,7 +430,7 @@ async function handleSearchProducts(args: Record<string, unknown>) {
                 region, country_code
          FROM products
          ORDER BY updated_at DESC
-         LIMIT $1`,
+         LIMIT $1::int`,
         [fetchLimit]
       );
       if (needsFilter) {
@@ -681,7 +681,7 @@ async function handleGetDeals(args: Record<string, unknown>) {
            AND price > 0
            AND country_code = $1
            AND search_vector @@ plainto_tsquery('english', $2)
-         LIMIT $3`,
+         LIMIT $3::int`,
         [effectiveCountry, fallbackQuery, Number(limit) || 20]
       );
       total = fallbackResult.rows.length;
@@ -1255,7 +1255,7 @@ async function handleFindSimilar(args: Record<string, unknown>) {
       `SELECT product_id, (embedding <=> $1::vector)::float AS distance
        FROM product_embeddings
        WHERE product_id != $2 AND model_ver = 'gemini-embedding-001@512'
-       ORDER BY distance LIMIT $3`,
+       ORDER BY distance LIMIT $3::int`,
       [refEmbedding, productId, limit]
     );
   } catch {
